Reduce code complexity in Adsense SettingsView.
